app/requirements.txt
streamlit
kaggle 
python-dotenv
duckdb